+++ test/example.c
@@ -5,17 +5,17 @@
 
 void ex1(void)
 {
-	int vlen = 0;
+	size_t vlen = 0;
 
 	char *json = "{\"foo\":\"bar\",\"barbar\":[1,2,3],\"obj\":{\"a\":\"b\"}}";
 	printf("parsing '%s'\n", json);
 	char *val = js0n("barbar", 0, json, strlen(json), &vlen);
-	printf("returned %.*s\n",vlen,val);
+	printf("returned %.*s\n",(int)vlen,val);
 
 	char *array = "[\"foo\",\"bar\",[1,2,3],{\"a\":\"b\"},42]";
 	printf("parsing '%s'\n", array);
 	val = js0n(0, 3, array, strlen(array), &vlen);
-	printf("returned %.*s\n",vlen,val);
+	printf("returned %.*s\n",(int)vlen,val);
 
 }
